refactor: reorg and renaming
