+++ src/background/crypto/generate-encryption-key.ts
