+++ src/background/crypto/generate-random-hex.ts
