+++ src/background/crypto/mnemonic-encryption.ts
