+++ src/background/popup.ts
